0.8.6.5 Upgrade calculations final update

+++ Hearts Of Greed/Motor2D/Hero.cpp
@@ -1457,6 +1457,7 @@ Skill::Skill(const Skill& skill1) : dmg(skill1.dmg), type(skill1.type), target(s
 
 
 //Getters and setters hellish nightmare
+// PD: Now I have to enter this nightmare as well. The quest for an efficient code has started
 
 int Hero::GetHeroLevel() const
 {
@@ -1578,13 +1579,13 @@ void Hero::SetFeelingSecure(float feelSecure)
 }
 
 
-int Hero::GetAttackDamage() const
+float Hero::GetAttackDamage() const
 {
 	return attackDamage;
 }
 
 
-void Hero::SetAttackDamage(int atkDamage)
+void Hero::SetAttackDamage(float atkDamage)
 {
 	attackDamage = atkDamage;
 }
